+++ apps/web/app/(use-page-wrapper)/settings/(settings-layout)/developer/webhooks/new/loading.tsx
@@ -1,4 +1,4 @@
-import { SkeletonLoader } from "@calcom/features/webhooks/pages/webhook-new-view";
+import { SkeletonLoader } from "@calcom/features/webhooks/pages/webhook-new-skeleton";
 
 export default function Loading() {
   return <SkeletonLoader />;
